tests/wpt/web-platform-tests/webstorage/event_session_url.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>WebStorage Test: sessionStorage event - url</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script>
      function fail(msg) {
          t.step(function() {
              assert_unreached(msg);
          });
          t.done();
      }
  </script>
 </head>
 <body>
    <h1>event_session_url</h1>
    <div id="log"></div>
    <script>
        var t = async_test("url property test of session event");

        test(function() {
            sessionStorage.clear();
            function onStorageEvent(event) {
                t.step(function() {
                    var url = window.location.href;

                    var pos = url.lastIndexOf("/");
                    if (pos != -1) {
                        url = url.substr(0, pos + 1);
                        url = url + "resources/session_set_item_iframe.html";
                    }

                   assert_equals(event.url, url);
                });
                t.done();
            }

            window.addEventListener('storage', onStorageEvent, false);

            var el = document.createElement("iframe");
            el.setAttribute('id', 'ifrm');
            el.setAttribute('src', 'resources/session_set_item_iframe.html');
            document.body.appendChild(el);
        }, "Session event is fired due to an invocation of the setItem() method.");
    </script>
 </body>
</html>
